+++ src/components/Team/components/TeamListHeader/index.tsx
@@ -6,6 +6,7 @@ import ArrowBackIcon from 'assets/svg/Team/arrow-back.svg';
 import NotificationIcon from 'assets/svg/Team/notification.svg';
 import ProfileIcon from 'assets/svg/Team/profile.svg';
 import ROUTES from 'static/routes';
+import useLogger from 'utils/hooks/analytics/useLogger';
 import useMount from 'utils/hooks/state/useMount';
 import useTokenState from 'utils/hooks/state/useTokenState';
 
@@ -15,6 +16,7 @@ export default function TeamListHeader() {
   const router = useRouter();
   const token = useTokenState();
   const isMounted = useMount();
+  const logger = useLogger();
 
   const { data: notificationData } = useQuery({
     ...teamQueries.notifications(token ?? '', { limit: 1 }),
@@ -32,6 +34,11 @@ export default function TeamListHeader() {
     router.back();
   };
 
+  const handleNotificationsClick = () => {
+    logger.actionEventClick({ team: 'CAMPUS', event_label: 'team_recruitment_notification', value: '알림' });
+    router.push(ROUTES.TeamNotifications());
+  };
+
   return (
     <header className={styles.header}>
       <button type="button" className={styles['header__back-button']} aria-label="뒤로가기" onClick={handleBack}>
@@ -45,7 +52,7 @@ export default function TeamListHeader() {
           type="button"
           className={styles['header__icon-button']}
           aria-label="알림"
-          onClick={() => router.push(ROUTES.TeamNotifications())}
+          onClick={handleNotificationsClick}
         >
           <NotificationIcon />
           {hasUnreadNotifications && <span className={styles['header__notification-dot']} />}
